add build release and clean all option

+++ scripts/ccomp.sh
@@ -5,7 +5,10 @@ if [[ "$1" == "clean" ]]; then
     cd $AGRELEASE/build
     cmake3 --build . --target clean
     cd $AGRELEASE
-    rm -rf $AGRELEASE/build $AGRELEASE/bin
+    if [[ "$2" == "all" ]]; then
+	echo "removing build and bin folders"
+	rm -rf $AGRELEASE/build $AGRELEASE/bin
+    fi
 
 elif [[ "$1" == "update" ]]; then
     echo "Recompiling agdaq"
@@ -56,6 +59,16 @@ elif [[ "$1" == "debug" ]]; then
     cd $AGRELEASE/build
 #
     cmake3 .. -DBUILD_AG_SIM=OFF -DBUILD_A2=OFF -DCMAKE_BUILD_TYPE=Debug
+    time cmake3 --build . -- -j
+    time cmake3 --build . --target install -- -j
+    cd $AGRELEASE
+
+elif [[ "$1" == "build" ]]; then
+    echo "Building agdaq optimized"
+    mkdir -p $AGRELEASE/build
+    cd $AGRELEASE/build
+#
+    cmake3 .. -DBUILD_AG_SIM=ON -DBUILD_A2=OFF -DCMAKE_BUILD_TYPE=Release
     time cmake3 --build . -- -j
     time cmake3 --build . --target install -- -j
     cd $AGRELEASE
